git: run `git-callback` using system ruby
When checking out a branch with a Ruby version in the `.ruby-version`
file that is not installed, the script will exit with an error saying
that Ruby is missing.

By using system Ruby instead of the one in the repository when running
the callback script, the script will keep on working even when the local
ruby version is missing.

The path to system Ruby might be different between platforms, so it
might make sense to rewrite this script as a shell script instead.

https://github.com/bzf/dotfiles/issues/12