zsh: extract out nvm stuff into nodejs tag
This was done in an attempt to try and figure out what slowed down my
shell startup times. Moving this around did nothing special, but it
became easier to understand it was `nvm` that caused most of the
slowdown.

Turns out that if you specify a `nvm default` version it slows down your
shell startup time.
Adding the `--no-use` flag to the `nvm.sh` call removes the slowdown but
doesn't allow you to specify your global nodejs version. I'm okay with
that tradeoff!

Without the flag:
```sh
$ time ~/.nvm/nvm.sh
~/.nvm/nvm.sh  0.46s user 0.14s system 101% cpu 0.597 total
```

With the flag:
```sh
$ time ~/.nvm/nvm.sh --no-use
~/.nvm/nvm.sh --no-use  0.02s user 0.00s system 94% cpu 0.024 total
```