vim: remove YouCompleteMe, snippets and disable Python
For whatever reason, having python enabled caused `neovim` to slow down
when having multiple splits open and a lot of jumping around.

After using the built in completion shortcuts I've found that using that
in the trade off of not having the slowdown while using the editor is
worth it.

I sort off miss having snippets but for now this is better.