Use :Files for <C-p> in vim
When trying to open a file that's not tracked it won't show up if you're
using the `:GFiles` command (since it's using `git ls-files` for listing
the available files).

This, together with 517a38bbf390df878bb2e58ec4805f73f8ccf190, will make
it possible to search _all_ files with a reasonable speed in the project
at work.