Add ale for running prettier in JS projects
Don't run it on save but instead map it to `<leader>f` so we can make
sure it doesn't run on files we don't want it to.

Would be nice if you could configure that differently per filetype (e.g.
in Elm I'd always want it to run on save). Probably doable, but it's
been a while since I wrote Elm so might be something to look into later.